Updated handling overlay for destroyed structures
Destroyed_Structure overlay graphics can now be smaller than 4x3 tiles of 16x16 pixels. If 4x3 is provided, destroyed structures will take a cut-out of the available graphics. In other words, if a 2x2 structure is destroyed that may lead to the top-left four tiles of the available graphics being used or the bottom-right ones or any other 2x2 set. If fewer tiles are made available for Destroyed_Structure, no such clever tactics are used. Instead, the available tiles are simply used consecutively.